==> 10_website_sce/06_SCDE_cluster.Rmd <==
---
title: "Differential expression between clusters using *scde*"
bibliography:
  bibtex.bib
---

<!-- As HTML pages are built independently, large objects need to be
reimported from disk, and smaller ones quickly recomputed. -->

```{r setup, include=FALSE}
knitr::opts_chunk$set(cache = TRUE)
stopifnot(
  require(scater),
  require(scran),
  require(ggplot2),
  require(ComplexHeatmap),
  require(dplyr),
  require(ggrepel),
  require(RColorBrewer),
  require(goseq),
  require(scde),
  requireNamespace("dplyr")
)
sce.endo <- readRDS("rds/sce.endo_clusters.rds")
o.ifm.2h <- readRDS("rds/o.ifm_2h.rds")
o.ifm.4h <- readRDS("rds/o.ifm_4h.rds")
o.ifm.6h <- readRDS("rds/o.ifm_6h.rds")
col.time <- brewer.pal(9, "Set3")[c(2,7:8)]
col.infection <- brewer.pal(12, "Paired")[c(9,4,2)]
col.status <- brewer.pal(12, "Paired")[c(9,1,7)]
names(col.time) <- levels(sce.endo$Time)[1:3]
names(col.infection) <- levels(sce.endo$Infection)
names(col.status) <- levels(sce.endo$Status)
outdir <- "06_SCDE_cluster_out"; dir.create(outdir, showWarnings = FALSE)
scde.res <- readRDS("rds/scde_cluster.rds")
goseq.res <- readRDS("rds/goseq_cluster_res.rds")
# Subset by time point
sce.2h <- sce.endo[,sce.endo$Time == '2h']
sce.4h <- sce.endo[,sce.endo$Time == '4h']
sce.6h <- sce.endo[,sce.endo$Time == '6h']
# Filter detected genes
filterCounts <- function(m, counts = 10, cells = 10){
  apply(m, 1, function(e){
    return(sum(e >= counts) >= cells)
  })
}
sg.2h <- droplevels(sce.2h$Group)
keep.2h <- filterCounts(counts(sce.2h))
cd.2h <- counts(sce.2h)[keep.2h,]; storage.mode(cd.2h) <- 'integer'
sg.4h <- droplevels(sce.4h$Group)
keep.4h <- filterCounts(counts(sce.4h))
cd.4h <- counts(sce.4h)[keep.4h,]; storage.mode(cd.4h) <- 'integer'
sg.6h <- droplevels(sce.6h$Group)
keep.6h <- filterCounts(counts(sce.6h))
cd.6h <- counts(sce.6h)[keep.6h,]; storage.mode(cd.6h) <- 'integer'
# Reorder SCESet to match count matrix
sce.ifm.2h <- sce.2h[,rownames(o.ifm.2h)]
cd.ifm.2h <- counts(sce.ifm.2h); storage.mode(cd.ifm.2h) <- "integer"
sce.ifm.4h <- sce.4h[,rownames(o.ifm.4h)]
cd.ifm.4h <- counts(sce.ifm.4h); storage.mode(cd.ifm.4h) <- "integer"
sce.ifm.6h <- sce.6h[,rownames(o.ifm.6h)]
cd.ifm.6h <- counts(sce.ifm.6h); storage.mode(cd.ifm.6h) <- "integer"
# Prior of expression
o.prior.2h <- scde.expression.prior(
  models = o.ifm.2h, counts = cd.ifm.2h, show.plot = FALSE
)
o.prior.4h <- scde.expression.prior(
  models = o.ifm.4h, counts = cd.ifm.4h, show.plot = FALSE
)
o.prior.6h <- scde.expression.prior(
  models = o.ifm.6h, counts = cd.ifm.6h, show.plot = FALSE
)
# Functions
convert.z.score <- function(x, one.sided = NULL) {
  z <- x$Z
  if(is.null(one.sided)) {
    pval = pnorm(-abs(z));
    pval = 2 * pval
  } else if(one.sided=="-") {
    pval = pnorm(z);
  } else {
    pval = pnorm(-z);
  }
    x <- cbind(
      x,
      p.value = pval
  )
  return(x);
}   
addGeneName <- function(x){
  x <- cbind(
    gene_name = with(rowData(sce.endo), gene_name[match(rownames(x), gene_id)]),
    x
  )
  return(x)
}
orderResults <- function(x){
  x <- x[with(x, order(abs(Z), decreasing = TRUE)),]
  return(x)
}
sig.levels <- c(0.05, 0.01)
volcano.sig <- data.frame(
  p.value = sig.levels,
  level = as.character(sig.levels)
)
volcano.mle <- function(x, sub = NULL){
  varName <- deparse(substitute(x))
  x <- convert.z.score(x)
  gg <- ggplot(x, aes(mle, -log10(p.value))) +
    geom_point(aes(colour = (cZ != 0))) +
    geom_hline(aes(yintercept=-log10(p.value),linetype=level), volcano.sig) +
    ggtitle(varName, sub)
  print(gg)
  return(x)
}
```

<!-- Cleaner code based on 05_SCDE
Remove ERCC spike-ins
!! Split cells by time
!! Retain genes detected at each time (10 counts, 5 cells, 1 group)
Identify cell clusters @ time
<!-- Error model: {cells @ time; genes @ time}
Prior @ time
DE: {group on all cells with NA; no batch correction; P-value < 0.01}
-->

# Preprocessed data

**Note:**
Similarly to the identification of [markers](06_markers.html) for each cluster,
this analysis uses data prepared identically to the
[differential expression between experimental groups using *scde*](05_SCDE.html):

* ERCC spike-in features are removed,
* Cells in the data set is subsetted by time point,
* Genes in the data set are subsetted for those detected at sufficient levels
  within each time point,
* Error models are computed on the subsetted data sets

In addition, this analysis diverges from the differential expression between
supervised experimental groups by consideration of
[unsupervised clusters identified within each time point](05_cluster.html).

# Differential expression {#DE}

## Setup

Let us first define a list to store the result tables returned by *scde*:

```{r scde.res_init, eval=FALSE}
scde.res <- list()
```

In addition, note that several functions defined in the first
[scde section](05_SCDE.html#DE) will be re-used here.

## Contrasts {.tabset}

### 2h (2 vs 3)

Identification of [unsupervised clusters](05_cluster.html) at 2h p.i. revealed
that cluster 2 and 3 to comprise
exclusively stimulated and majoritarily uninfected DCs, respectively.
However, both clusters display some degree of overlap between their
[markers genes](06_markers.html).
Thus, a direct comparison of the two unsupervised clusters may detect more
subtle differences and effect sizes:

```{r contrasts.2h_2v1, eval=FALSE}
groupTarget <- 2; groupRef <- 3
sg.test <- factor(sce.ifm.2h$quickCluster.2h, levels = c(groupTarget, groupRef))
names(sg.test) <- colnames(sce.ifm.2h); summary(sg.test)
contrastName <- sprintf("2h_cluster%s-cluster%s", groupTarget, groupRef); message(contrastName)
scde.res[[contrastName]] <- scde.expression.difference(
    o.ifm.2h, cd.2h, o.prior.2h, sg.test, n.cores = 4, verbose = 1
  )
```

### 2h (2 vs 1)

In addition, cluster 2 (exclusively stimulated) may also be compared directly
to cluster 1, the latter displaying a large and even representation of
uninfected and challenged cells.
Again, a direct comparison of the two unsupervised clusters may detect
subtle differences and effect sizes:

```{r contrasts.2h_2v3, eval=FALSE}
groupTarget <- 2; groupRef <- 1
sg.test <- factor(sce.ifm.2h$quickCluster.2h, levels = c(groupTarget, groupRef))
names(sg.test) <- colnames(sce.ifm.2h); summary(sg.test)
contrastName <- sprintf("2h_cluster%s-cluster%s", groupTarget, groupRef); message(contrastName)
scde.res[[contrastName]] <- scde.expression.difference(
    o.ifm.2h, cd.2h, o.prior.2h, sg.test, n.cores = 4, verbose = 1
  )
```

### 4h (2 vs 1)

At 4h p.i., [unsupervised clusters](05_cluster.html) reveals a marked
distinction between _exposed_ and _infected_ DCs, accounting for the majority
of clusters 2 and 1, respectively, irrespective of the infection.
Thus, a direct comparison of the two unsupervised clusters may detect more
subtle differences and effect sizes:

```{r contrasts.4h_2v1, eval=FALSE}
groupTarget <- 2; groupRef <- 1
sg.test <- factor(sce.ifm.4h$quickCluster.4h, levels = c(groupTarget, groupRef))
names(sg.test) <- colnames(sce.ifm.4h); summary(sg.test)
contrastName <- sprintf("4h_cluster%s-cluster%s", groupTarget, groupRef); message(contrastName)
scde.res[[contrastName]] <- scde.expression.difference(
    o.ifm.4h, cd.4h, o.prior.4h, sg.test, n.cores = 4, verbose = 1
  )
```

### 6h (2 vs 1)

At 4h p.i., [unsupervised clusters](05_cluster.html) reveals a notable 
under-representation of _STM-D23580 exposed_ DCs in cluster 1 relative to cluster 2,
in contrast to _STM-LT2 exposed_ DCs showing the opposite trend.
Thus, a direct comparison of the two unsupervised clusters may detect more
subtle differences and effect sizes:

```{r contrasts.6h_run, eval=FALSE}
groupTarget <- 2; groupRef <- 1
sg.test <- factor(sce.ifm.6h$quickCluster.6h, levels = c(groupTarget, groupRef))
names(sg.test) <- colnames(sce.ifm.6h); summary(sg.test)
contrastName <- sprintf("6h_cluster%s-cluster%s", groupTarget, groupRef); message(contrastName)
scde.res[[contrastName]] <- scde.expression.difference(
    o.ifm.6h, cd.6h, o.prior.6h, sg.test, n.cores = 4, verbose = 1
  )
```

```{r scde.table_csv, echo=FALSE, eval=FALSE}
saveRDS(scde.res, "rds/scde_cluster.rds")
for (contrastName in names(scde.res)){
  scde.table <- scde.res[[contrastName]]
  scde.table <- convert.z.score(addGeneName(orderResults(scde.table)))
  csv.file <- sprintf("SCDE_cluster_%s.csv", contrastName)
  write.csv(scde.table, file.path(outdir, csv.file))
}
```

# Volcano plots {.tabset}

Let us first identify the extreme values of maximum likelihood estimate of
fold-change to scale subsequent plot for comparability:

```{r mleRange}
mleRange <-
  max(abs(do.call("c", lapply(scde.res, function(x){return(x$mle)}))))*c(-1,1)
pRange <- c(
    0,
    -log10(min(do.call("c", lapply(scde.res, function(x){
      return(convert.z.score(x)$p.value)
    }))))
)
```

## 2h (2 vs 1)

```{r volcano_2h_2v1, echo=FALSE}
v.data.2h <- scde.res[["2h_cluster2-cluster1"]][,c("mle", "Z", "cZ")]
v.data.2h <- convert.z.score(v.data.2h)
ggplot(v.data.2h) +
  geom_point(aes(mle, -log10(p.value), colour = (cZ != 0)), size = 0.5) +
  geom_hline(
    aes(yintercept = -log10(p.value), linetype = level), data = volcano.sig
  ) +
  scale_x_continuous(limits = mleRange) +
  scale_y_continuous(limits = pRange) +
  theme_bw() +
  labs(title = "2h: cluster 2 - cluster 1")
```

```{r fig1_volcano_2h_2v1, include=FALSE}
geneData.2h <- subset(addGeneName(v.data.2h),
  gene_name %in% c(
    "NFKB1","IRF8","STAT5A","CEBPA","CD1A","CD1C","CD1E","CLECL1", # 8
    "IL1B","IL1A","EBI3","CD40","CD83", # 5
    "CTSS","CTSD","CTSL","GRN","APOE","APOC1","ACOT2", # 7
    "S100A9","STAB1","LRP1","PSAP") # 4
)
ggplot(v.data.2h) +
  geom_point(aes(Z, -log10(p.value), colour = p.value < 0.01), size = 0.5) +
  geom_hline(yintercept = -log10(0.01), linetype = "dashed", alpha = 0.25) +
  geom_text_repel(
    aes(Z, -log10(p.value), label = gene_name), geneData.2h,
    size = 2, min.segment.length = unit(0, "mm"), alpha = 0.75,
    nudge_x = ifelse(geneData.2h$Z > 0, 1, -1), nudge_y = 0, segment.size = 0.2,
    fontface = "bold"
  ) +
  scale_x_continuous(limits = mleRange) +
  scale_y_continuous(limits = pRange) +
  theme_bw() +
  labs(
    title = "2h: cluster 2 - cluster 1",
    y = expression(-log[10]*" ("*italic(P)*"-value)")
  )
ggsave(file.path(outdir, "2h_volcano_Vogel.pdf"), height = 5, width = 6.5)
```

```{r fig2_volcano_2h_2v1, echo=FALSE, eval=FALSE}
ggplot(v.data.2h) +
  geom_point(
    aes(mle, -log10(p.value), colour = p.value < 0.01),
    size = 0.5, alpha = 0.25
  ) +
  geom_hline(yintercept = -log10(0.01), linetype = "dashed") +
  geom_text_repel(
    aes(mle, -log10(p.value), label = gene_name), geneData.2h,
    size = 2, min.segment.length = unit(0, "mm"), fontface = "bold"
  ) +
  scale_x_continuous(limits = mleRange) +
  scale_y_continuous(limits = pRange) +
  theme_bw() +
  labs(title = "2h: cluster 2 - cluster 1")
ggsave(file.path(outdir, "2h_volcano_Shalek.pdf"), height = 5, width = 6.5)
```

## 4h (2 vs 1)

```{r volcano_4h_2v1, echo=FALSE}
v.data.4h <- scde.res[["4h_cluster2-cluster1"]][,c("mle", "Z", "cZ")]
v.data.4h <- convert.z.score(v.data.4h)
ggplot(v.data.4h) +
  geom_point(aes(mle, -log10(p.value), colour = (cZ != 0)), size = 0.5) +
  geom_hline(
    aes(yintercept = -log10(p.value), linetype = level), data = volcano.sig
  ) +
  scale_x_continuous(limits = mleRange) +
  scale_y_continuous(limits = pRange) +
  theme_bw() +
  labs(title = "4h: cluster 2 - cluster 1")
```

```{r fig1_volcano_4h_2v1, include=FALSE}
geneData.4h <- subset(addGeneName(v.data.4h),
  gene_name %in% c(
    "NFKB1","IRF8","STAT5A","CEBPA","CD1A","CD1C","CD1E","CLECL1", # 8
    "IL1B","IL1A","EBI3","CD40","CD83", # 5
    "CTSS","CTSD","CTSL","GRN","APOE","APOC1","ACOT2", # 7
    "S100A9","STAB1","LRP1","PSAP") # 4
)
ggplot(v.data.4h) +
  geom_point(aes(Z, -log10(p.value), colour = p.value < 0.01), size = 0.5) +
  geom_hline(yintercept = -log10(0.01), linetype = "dashed", alpha = 0.25) +
  geom_text_repel(
    aes(Z, -log10(p.value), label = gene_name), geneData.4h,
    size = 2, min.segment.length = unit(0, "mm"), alpha = 0.75,
    nudge_x = ifelse(geneData.4h$Z > 0, 1, -1), nudge_y = 0, segment.size = 0.2,
    fontface = "bold"
  ) +
  scale_x_continuous(limits = mleRange) +
  scale_y_continuous(limits = pRange) +
  theme_bw() +
  labs(
    title = "4h: cluster 2 - cluster 1",
    y = expression(-log[10]*" ("*italic(P)*"-value)")
  )
ggsave(file.path(outdir, "4h_volcano_Vogel.pdf"), height = 5, width = 6.5)
```

```{r fig2_volcano_4h_2v1, echo=FALSE, eval=FALSE}
ggplot(v.data.4h) +
  geom_point(
    aes(mle, -log10(p.value), colour = p.value < 0.01),
    size = 0.5, alpha = 0.25
  ) +
  geom_hline(yintercept = -log10(0.01), linetype = "dashed") +
  geom_text_repel(
    aes(mle, -log10(p.value), label = gene_name), geneData.4h,
    size = 2, min.segment.length = unit(0, "mm"), fontface = "bold"
  ) +
  scale_x_continuous(limits = mleRange) +
  scale_y_continuous(limits = pRange) +
  theme_bw() +
  labs(title = "4h: cluster 2 - cluster 1")
ggsave(file.path(outdir, "4h_volcano_Shalek.pdf"), height = 5, width = 6.5)
```

## 6h (2 vs 1)

```{r volcano_6h_2v1, echo=FALSE}
v.data.6h <- scde.res[["6h_cluster2-cluster1"]][,c("mle", "Z", "cZ")]
v.data.6h <- convert.z.score(v.data.6h)
ggplot(v.data.6h) +
  geom_point(aes(mle, -log10(p.value), colour = (cZ != 0)), size = 0.5) +
  geom_hline(
    aes(yintercept = -log10(p.value), linetype = level), data = volcano.sig
  ) +
  scale_x_continuous(limits = mleRange) +
  scale_y_continuous(limits = pRange) +
  theme_bw() +
  labs(title = "6h: cluster 2 - cluster 1")
```

```{r fig1_volcano_6h_2v1, echo=FALSE, eval=FALSE}
geneData.6h <- subset(addGeneName(v.data.6h),
  gene_name %in% c(
    "CTSB","CTSC","CTSD","CTSL", # 4
    "DEFB1","DUOX","ATF3", # 3
    "IL12B","IL23A","IL1A","EBI3","IL6", # 5
    "MAFF","CREBFR","NFKB1","NFKB2",
    "MARCH1") # 4
)
ggplot(v.data.6h) +
  geom_point(aes(Z, -log10(p.value), colour = p.value < 0.01), size = 0.5) +
  geom_hline(yintercept = -log10(0.01), linetype = "dashed", alpha = 0.25) +
  geom_text_repel(
    aes(Z, -log10(p.value), label = gene_name), geneData.6h,
    size = 2, min.segment.length = unit(0, "mm"), alpha = 0.75,
    nudge_x = ifelse(geneData.6h$Z > 0, 1, -1), nudge_y = 0, segment.size = 0.2,
    fontface = "bold"
  ) +
  scale_x_continuous(limits = mleRange) +
  scale_y_continuous(limits = pRange) +
  theme_bw() +
  labs(
    title = "6h: cluster 2 - cluster 1",
    y = expression(-log[10]*" ("*italic(P)*"-value)")
  )
ggsave(file.path(outdir, "6h_volcano_Vogel.pdf"), height = 5, width = 6.5)
```

```{r fig2_volcano_6h_2v1, echo=FALSE, eval=FALSE}
ggplot(v.data.6h) +
  geom_point(
    aes(mle, -log10(p.value), colour = p.value < 0.01),
    size = 0.5, alpha = 0.25
  ) +
  geom_hline(yintercept = -log10(0.01), linetype = "dashed") +
  geom_text_repel(
    aes(mle, -log10(p.value), label = gene_name), geneData.6h,
    size = 2, min.segment.length = unit(0, "mm"), fontface = "bold"
  ) +
  scale_x_continuous(limits = mleRange) +
  scale_y_continuous(limits = pRange) +
  theme_bw() +
  labs(title = "6h: cluster 2 - cluster 1")
ggsave(file.path(outdir, "6h_volcano_Shalek.pdf"), height = 5, width = 6.5)
```

# Count DE genes at various cut-offs {.tabset}

## 2h (2 vs 1)

```{r countDE_2h_2v1, echo=FALSE, results='asis'}
v.data <- scde.res[["2h_cluster2-cluster1"]][,c("Z", "cZ")]
v.data <- convert.z.score(v.data)
v.data <- data.frame(
  P.01 = sum(v.data$p.value < 0.01),
  P.05 = sum(v.data$p.value < 0.05),
  cZ = sum(v.data$cZ != 0),
  row.names = "2h: cluster2 - cluster1"
)
pander::pandoc.table(v.data)
```

## 4h (2 vs 1)

```{r countDE_4h_2v1, echo=FALSE, results='asis'}
v.data <- scde.res[["4h_cluster2-cluster1"]][,c("Z", "cZ")]
v.data <- convert.z.score(v.data)
v.data <- data.frame(
  P.01 = sum(v.data$p.value < 0.01),
  P.05 = sum(v.data$p.value < 0.05),
  cZ = sum(v.data$cZ != 0),
  row.names = "4h: cluster2 - cluster1"
)
pander::pandoc.table(v.data)
```

## 6h (2 vs 1)

```{r countDE_6h_2v1, echo=FALSE, results='asis'}
v.data <- scde.res[["6h_cluster2-cluster1"]][,c("Z", "cZ")]
v.data <- convert.z.score(v.data)
v.data <- data.frame(
  P.01 = sum(v.data$p.value < 0.01),
  P.05 = sum(v.data$p.value < 0.05),
  cZ = sum(v.data$cZ != 0),
  row.names = "6h: cluster2 - cluster1"
)
pander::pandoc.table(v.data)
```

# Gene ontology

Let us use the [goseq](http://bioconductor.org/packages/goseq) package to
identify the most enriched gene ontologies among the various lists of DE genes.
Note that we restrict the results to GO categories associated with at least
**10** genes, for robustness.

Let us first define:

* a list to store the GO enrichment tables returned by *goseq*

```{r goseq.res_init, eval=FALSE}
goseq.res <- list()
```

* the gene length information reported by *featureCounts*:

```{r geneLengths}
geneLengths <- width(sce.endo)
names(geneLengths) <- rownames(sce.endo)
```

Let us identify over- and under-represented GO categories in DE genes
identified in each contrast,
using as background the list of genes detected at each time point:

## Estimate enrichment {.tabset}

### 2h (2 vs 1)

```{r GO_2h_DE, eval=FALSE}
contrastName <- "2h_cluster2-cluster1"
bg.2h <- rownames(sce.ifm.2h)
scde.table <- scde.res[[contrastName]]
scde.table <- convert.z.score(scde.table)
scde.table <- subset(scde.table, p.value < 0.01)
de.genes <- (bg.2h %in% rownames(scde.table))
names(de.genes) <- bg.2h; table(de.genes)
pwf <- nullp(de.genes, bias.data = geneLengths[names(de.genes)])
go.res <- goseq(pwf, "hg38", "ensGene")
goseq.res[[contrastName]] <- go.res
```

### 4h (2 vs 1)

```{r GO_4h_DE, eval=FALSE}
contrastName <- "4h_cluster2-cluster1"
bg.4h <- rownames(sce.ifm.4h)
scde.table <- scde.res[[contrastName]]
scde.table <- convert.z.score(scde.table)
scde.table <- subset(scde.table, p.value < 0.01)
de.genes <- (bg.4h %in% rownames(scde.table))
names(de.genes) <- bg.4h; table(de.genes)
pwf <- nullp(de.genes, bias.data = geneLengths[names(de.genes)])
go.res <- goseq(pwf, "hg38", "ensGene")
goseq.res[[contrastName]] <- go.res
```

### 6h (2 vs 1)

```{r GO_6h_DE, eval=FALSE}
contrastName <- "6h_cluster2-cluster1"
bg.6h <- rownames(sce.ifm.6h)
scde.table <- scde.res[[contrastName]]
scde.table <- convert.z.score(scde.table)
scde.table <- subset(scde.table, p.value < 0.01)
de.genes <- (bg.6h %in% rownames(scde.table))
names(de.genes) <- bg.6h; table(de.genes)
pwf <- nullp(de.genes, bias.data = geneLengths[names(de.genes)])
go.res <- goseq(pwf, "hg38", "ensGene")
goseq.res[[contrastName]] <- go.res
```

```{r goseq.res, echo=FALSE, eval=FALSE}
saveRDS(goseq.res, "rds/goseq_cluster_res.rds")
for (contrastName in names(goseq.res)){
  goseq.table <- goseq.res[[contrastName]]
  csvFile <- sprintf("SCDE_cluster_direct_GO_%s.csv", contrastName)
  write.csv(goseq.table, file.path(outdir, csvFile))
}
```
