==> 10_website_sce/09_markers.Rmd <==
---
title: "Markers of clusters using *scde*"
bibliography:
  bibtex.bib
editor_options: 
  chunk_output_type: console
---

<!-- As HTML pages are built independently, large objects need to be
reimported from disk, and smaller ones quickly recomputed. -->

```{r setup, include=FALSE}
knitr::opts_chunk$set(cache = TRUE)
stopifnot(
  require(scater),
  require(scran),
  require(ggplot2),
  require(ComplexHeatmap),
  require(dplyr),
  require(ggrepel),
  require(RColorBrewer),
  require(goseq),
  require(scde),
  requireNamespace("circlize")
)
sce.endo <- readRDS("rds/sce.endo_clusters.rds")
o.ifm.2h <- readRDS("rds/o.ifm.2h.rds")
o.ifm.4h <- readRDS("rds/o.ifm.4h.rds")
o.ifm.6h <- readRDS("rds/o.ifm.6h.rds")
col.time <- brewer.pal(9, "Set3")[c(2,7:8)]
col.infection <- brewer.pal(12, "Paired")[c(9,4,2)]
col.status <- brewer.pal(12, "Paired")[c(9,1,7)]
names(col.time) <- levels(sce.endo$Time)[1:3]
names(col.infection) <- levels(sce.endo$Infection)
names(col.status) <- levels(sce.endo$Status)
outdir <- "09_markers_out"; dir.create(outdir, showWarnings = FALSE)
scde.res <- readRDS("rds/scde_markers.rds")
goseq.res <- readRDS("rds/goseq_markers.rds")
# Subset by time point
sce.2h <- sce.endo[,sce.endo$Time == '2h']
sce.4h <- sce.endo[,sce.endo$Time == '4h']
sce.6h <- sce.endo[,sce.endo$Time == '6h']
# Filter detected genes
filterCounts <- function(m, counts = 10, cells = 10){
  apply(m, 1, function(e){
    return(sum(e >= counts) >= cells)
  })
}
# Reorder SCESet to match count matrix
sce.ifm.2h <- sce.2h[,rownames(o.ifm.2h)]
sg.ifm.2h <- droplevels(sce.ifm.2h$Group)
keep.2h <- filterCounts(counts(sce.ifm.2h))
cd.ifm.2h <- counts(sce.ifm.2h)[keep.2h,]; storage.mode(cd.ifm.2h) <- 'integer'
sce.ifm.4h <- sce.4h[,rownames(o.ifm.4h)]
sg.ifm.4h <- droplevels(sce.ifm.4h$Group)
keep.4h <- filterCounts(counts(sce.ifm.4h))
cd.ifm.4h <- counts(sce.ifm.4h)[keep.4h,]; storage.mode(cd.ifm.4h) <- 'integer'
sce.ifm.6h <- sce.6h[,rownames(o.ifm.6h)]
sg.ifm.6h <- droplevels(sce.ifm.6h$Group)
keep.6h <- filterCounts(counts(sce.ifm.6h))
cd.ifm.6h <- counts(sce.ifm.6h)[keep.6h,]; storage.mode(cd.ifm.6h) <- 'integer'
# Prior of expression
o.prior.2h <- scde.expression.prior(
  models = o.ifm.2h, counts = cd.ifm.2h, show.plot = FALSE
)
o.prior.4h <- scde.expression.prior(
  models = o.ifm.4h, counts = cd.ifm.4h, show.plot = FALSE
)
o.prior.6h <- scde.expression.prior(
  models = o.ifm.6h, counts = cd.ifm.6h, show.plot = FALSE
)
# Functions
convert.z.score <- function(x, one.sided = NULL) {
  z <- x$Z
  if(is.null(one.sided)) {
    pval = pnorm(-abs(z));
    pval = 2 * pval
  } else if(one.sided=="-") {
    pval = pnorm(z);
  } else {
    pval = pnorm(-z);
  }
    x <- cbind(
      x,
      p.value = pval
  )
  return(x);
}   
addGeneName <- function(x){
  x <- cbind(
    gene_name = with(rowData(sce.endo), gene_name[match(rownames(x), gene_id)]),
    x
  )
  return(x)
}
orderResults <- function(x){
  x <- x[with(x, order(abs(Z), decreasing = TRUE)),]
  return(x)
}
sig.levels <- c(0.05, 0.01)
volcano.sig <- data.frame(
  P = sig.levels,
  level = as.character(sig.levels)
)
volcano.mle <- function(x, sub = NULL){
  varName <- deparse(substitute(x))
  x <- convert.z.score(x)
  gg <- ggplot(x, aes(mle, -log10(p.value))) +
    geom_point(aes(colour = (cZ != 0))) +
    geom_hline(aes(yintercept=-log10(p.value),linetype=level),volcano.sig) +
    ggtitle(varName, sub)
  print(gg)
  return(x)
}
```

<!-- Cleaner code based on 05_SCDE
Remove ERCC spike-ins
!! Split cells by time
!! Retain genes detected at each time (10 counts, 5 cells, 1 group)
Identify cell clusters @ time
<!-- Error model: {cells @ time; genes @ time}
Prior @ time
DE: {group on all cells with NA; no batch correction; P-value < 0.01}
-->

# Preprocessed data

**Note:**
for this analysis, the data is prepared identically to the
[differential expression between experimental groups using *scde*](05_SCDE.html):

* ERCC spike-in features are removed,
* Cells in the data set is subsetted by time point,
* Genes in the data set are subsetted for those detected at sufficient levels
  within each time point,
* Error models are computed on the subsetted data sets

# Differential expression {#DE .tabset}

As a preliminary step to the identification of marker genes for each cluster,
let us perform differential expression analysis between each cluster
compared against the union of all other cells within that time point.

## Setup

Let us first define a list to store the result tables returned by *scde*:

```{r scde.res_init, eval=FALSE}
scde.res <- list()
```

In addition, note that several functions defined in the first
[scde section](08_SCDE_v8.html#DE) will be re-used here.

## 2h

```{r scde.res_2h, eval=FALSE}
stopifnot(all(rownames(o.ifm.2h) == colnames(cd.ifm.2h)))
for (clusterId in levels(sce.ifm.2h$quickCluster.2h)){
  sg.test <- factor(
    sce.ifm.2h$quickCluster.2h == clusterId,
    levels = c(TRUE, FALSE),
    labels = c(clusterId, "Others"))
  names(sg.test) <- colnames(sce.ifm.2h); summary(sg.test)
  contrastName <- sprintf("2h_cluster%s-Others", clusterId);message(contrastName)
  scde.res[[contrastName]] <- scde.expression.difference(
      o.ifm.2h, cd.ifm.2h, o.prior.2h, sg.test, n.cores = 4, verbose = 1
    )
}
```

## 4h

```{r scde.res_4h, eval=FALSE}
stopifnot(all(rownames(o.ifm.4h) == colnames(cd.ifm.4h)))
for (clusterId in levels(sce.ifm.4h$quickCluster.4h)){
  sg.test <- factor(
    sce.ifm.4h$quickCluster.4h == clusterId,
    levels = c(TRUE, FALSE),
    labels = c(clusterId, "Others"))
  names(sg.test) <- colnames(sce.ifm.4h); summary(sg.test)
  contrastName <- sprintf("4h_cluster%s-Others", clusterId);message(contrastName)
  scde.res[[contrastName]] <- scde.expression.difference(
      o.ifm.4h, cd.ifm.4h, o.prior.4h, sg.test, n.cores = 4, verbose = 1
    )
}
```

## 6h

```{r scde.res_6h, eval=FALSE}
stopifnot(all(rownames(o.ifm.6h) == colnames(cd.ifm.6h)))
for (clusterId in levels(sce.ifm.6h$quickCluster.6h)){
  sg.test <- factor(
    sce.ifm.6h$quickCluster.6h == clusterId,
    levels = c(TRUE, FALSE),
    labels = c(clusterId, "Others"))
  names(sg.test) <- colnames(sce.ifm.6h); summary(sg.test)
  contrastName <- sprintf("6h_cluster%s-Others", clusterId);message(contrastName)
  scde.res[[contrastName]] <- scde.expression.difference(
      o.ifm.6h, cd.ifm.6h, o.prior.6h, sg.test, n.cores = 4, verbose = 1
    )
}
```

```{r scde.table_csv, echo=FALSE, eval=FALSE}
saveRDS(scde.res, "rds/scde_markers.rds")
for (contrastName in names(scde.res)){
  scde.table <- scde.res[[contrastName]]
  scde.table <- convert.z.score(addGeneName(orderResults(scde.table)))
  csv.file <- sprintf("SCDE_markers_%s.csv", contrastName)
  write.csv(scde.table, file.path(outdir, csv.file))
}
```

# Top 50 markers by fold-change

## Identification {.tabset}

For each time point, let us identify up to 50 gene markers for each cluster,
and trace the respective cluster for which each gene is a marker,
selected as follows:

* *P*-value less than 0.01
* *mle* of log fold-change among the 50 largest positive values for a cluster

**Note**:
a gene may be identified as a marker for more than one cluster.

### 2h

```{r markers_2h_FC}
markers_2h_FC <- character()
markers_2h_FC.cluster <- factor(NULL, levels(sce.ifm.2h$quickCluster.2h))
c2h <- grep("^2h", names(scde.res), value = TRUE)
for (contrastName in c2h){
  clusterId <- gsub("2h_cluster(.)-Others", "\\1", contrastName)
  scde.table <- addGeneName(convert.z.score(scde.res[[contrastName]]))
  scde.table <- subset(scde.table, mle > 0 & p.value < 0.01)
  scde.table <- scde.table[order(scde.table$mle, decreasing = TRUE),]
  scde.markers <- head(rownames(scde.table), 50)
  markers_2h_FC <- c(markers_2h_FC, scde.markers)
  markers_2h_FC.cluster <-
    c(markers_2h_FC.cluster, rep(clusterId, length(scde.markers)))
}
```

### 4h

```{r markers_4h_FC}
markers_4h_FC <- character()
markers_4h_FC.cluster <- factor(NULL, levels(sce.ifm.4h$quickCluster.4h))
c4h <- grep("^4h", names(scde.res), value = TRUE)
for (contrastName in c4h){
  clusterId <- gsub("4h_cluster(.)-Others", "\\1", contrastName)
  scde.table <- addGeneName(convert.z.score(scde.res[[contrastName]]))
  scde.table <- subset(scde.table, mle > 0 & p.value < 0.01)
  scde.table <- scde.table[order(scde.table$mle, decreasing = TRUE),]
  scde.markers <- head(rownames(scde.table), 50)
  markers_4h_FC <- c(markers_4h_FC, scde.markers)
  markers_4h_FC.cluster <-
    c(markers_4h_FC.cluster, rep(clusterId, length(scde.markers)))
}
```

### 6h

```{r markers_6h_FC}
markers_6h_FC <- character()
markers_6h_FC.cluster <- factor(NULL, levels(sce.ifm.6h$quickCluster.6h))
c6h <- grep("^6h", names(scde.res), value = TRUE)
for (contrastName in c6h){
  clusterId <- gsub("6h_cluster(.)-Others", "\\1", contrastName)
  scde.table <- addGeneName(convert.z.score(scde.res[[contrastName]]))
  scde.table <- subset(scde.table, mle > 0 & p.value < 0.01)
  scde.table <- scde.table[order(scde.table$mle, decreasing = TRUE),]
  scde.markers <- head(rownames(scde.table), 50)
  markers_6h_FC <- c(markers_6h_FC, scde.markers)
  markers_6h_FC.cluster <-
    c(markers_6h_FC.cluster, rep(clusterId, length(scde.markers)))
}
```

## Heat map of fold-change for (up to) 50 markers {.tabset}

### 2h

```{r markers_FC_2h, echo=FALSE, fig.height=10, message=FALSE}
markersMatrix <- matrix(
  nrow = length(markers_2h_FC),
  ncol = length(c2h),
  dimnames = list(
    gene_id = markers_2h_FC,
    cluster = gsub("2h_cluster(.)-Others", "\\1", c2h)
  ))
for (contrastName in c2h){
  scde.table <- addGeneName(convert.z.score(scde.res[[contrastName]]))
  clusterName <- gsub("2h_cluster(.)-Others", "\\1", contrastName)
  markersMatrix[,clusterName] <- scde.table[rownames(markersMatrix),"mle"]
}
rownames(markersMatrix) <-
  with(rowData(sce.2h), gene_name[match(rownames(markersMatrix), gene_id)])
ht <- Heatmap(
  matrix = markersMatrix,
  cluster_rows = FALSE, cluster_columns = FALSE,
  row_names_gp = gpar(cex=0.4)
  )
draw(ht)
```

```{r markers_FC_2h_pdf, include=FALSE}
pdf(file.path(outdir, "2h_markers_FC_heatmap_FC.pdf"), width = 8, height = 15)
draw(ht)
dev.off()
```

### 4h

```{r markers_FC_4h, echo=FALSE, fig.height=10}
markersMatrix <- matrix(
  nrow = length(markers_4h_FC),
  ncol = length(c4h),
  dimnames = list(
    gene_id = markers_4h_FC,
    cluster = gsub("4h_cluster(.)-Others", "\\1", c4h)
  ))
for (contrastName in c4h){
  scde.table <- addGeneName(convert.z.score(scde.res[[contrastName]]))
  clusterName <- gsub("4h_cluster(.)-Others", "\\1", contrastName)
  markersMatrix[,clusterName] <- scde.table[rownames(markersMatrix),"mle"]
}
rownames(markersMatrix) <-
  with(rowData(sce.4h), gene_name[match(rownames(markersMatrix), gene_id)])
ht <- Heatmap(
  matrix = markersMatrix,
  cluster_rows = FALSE, cluster_columns = FALSE,
  row_names_gp = gpar(cex=0.4)
  )
draw(ht)
```

```{r markers_FC_4h_pdf, include=FALSE}
pdf(file.path(outdir, "4h_markers_FC_heatmap_FC.pdf"), width = 8, height = 15)
draw(ht)
dev.off()
```

### 6h

```{r markers_FC_6h, echo=FALSE, fig.height=10}
markersMatrix <- matrix(
  nrow = length(markers_6h_FC),
  ncol = length(c6h),
  dimnames = list(
    gene_id = markers_6h_FC,
    cluster = gsub("6h_cluster(.)-Others", "\\1", c6h)
  ))
for (contrastName in c6h){
  scde.table <- addGeneName(convert.z.score(scde.res[[contrastName]]))
  clusterName <- gsub("6h_cluster(.)-Others", "\\1", contrastName)
  markersMatrix[,clusterName] <- scde.table[rownames(markersMatrix),"mle"]
}
rownames(markersMatrix) <-
  with(rowData(sce.6h), gene_name[match(rownames(markersMatrix), gene_id)])
ht <- Heatmap(
  matrix = markersMatrix,
  cluster_rows = FALSE, cluster_columns = FALSE,
  row_names_gp = gpar(cex=0.4)
  )
draw(ht)
```

```{r markers_FC_6h_pdf, include=FALSE}
pdf(file.path(outdir, "6h_markers_FC_heatmap_FC.pdf"), width = 8, height = 15)
draw(ht)
dev.off()
```

## Heat map of row-scaled log-counts for (up to) 50 markers {.tabset}

Notably, let us use here the same markers as defined in the previous section,
namely:

* *P*-value less than 0.01
* *mle* of log fold-change among the 50 largest positive values for a cluster

First of all, let us define a color scale share between upcoming 
row-scaled heat maps:

```{r ht.col}
ht.col <- c("purple", "black", "yellow")
```

### 2h

```{r markers_FC_logcounts_2h, echo=FALSE, fig.height=10, message=FALSE}
ht_list <- list()
markersMatrix <- assay(sce.ifm.2h, "logcounts")[markers_2h_FC,]
markersMatrix <- t(scale(t(markersMatrix)))
rownames(markersMatrix) <- with(
    rowData(sce.2h), gene_name[match(rownames(markersMatrix), gene_id)])
# maxMatrix <- max(abs(range(markersMatrix))) * ht.scale.percentile
ht_scale <- circlize::colorRamp2(c(-4, 0, 4), ht.col) # use maxMatrix?
for (clusterId in levels(sce.ifm.2h$quickCluster.2h)){
  clusterMatrix <- markersMatrix[,sce.ifm.2h$quickCluster.2h == clusterId]
  ha_tmp <- HeatmapAnnotation(
    df = data.frame(subset(
      colData(sce.ifm.2h), quickCluster.2h == clusterId,
      c("Infection", "Status")
    )),
    col = list(
      Infection = col.infection,
      Status = col.status
    )
  )
  ht_list[[clusterId]] <- Heatmap(
    matrix = clusterMatrix,
    name = ifelse(
      clusterId == levels(sce.ifm.2h$quickCluster.2h)[1],
      "z-score",
      clusterId),
    col = ht_scale,
    row_names_gp = gpar(cex=0.4),
    show_heatmap_legend =
      clusterId == levels(sce.ifm.2h$quickCluster.2h)[1],
    show_row_names = clusterId == levels(sce.ifm.2h$quickCluster.2h)[1],
    cluster_rows = FALSE,
    show_column_names = FALSE, show_column_dend = FALSE,
    row_names_side = "left",
    top_annotation = ha_tmp,
    split = markers_2h_FC.cluster
  )
}
ht_final <- ht_list[["0"]] + ht_list[["1"]] + ht_list[["2"]] + ht_list[["3"]]
draw(ht_final)
```

```{r markers_FC_logcounts_2h_pdf, include=FALSE}
pdf(
  file.path(outdir, "2h_markers_FC_heatmap_logcounts.pdf"),
  width = 10, height = 15)
draw(ht_final)
dev.off()
```

### 2h (w/out '0')

```{r markers_FC_logcounts_2h_no0, echo=FALSE, fig.height=10, message=FALSE}
ht_list <- list()
excludeMarkers <- which(markers_2h_FC.cluster == "0")
markers_2h_FC_no0 <- markers_2h_FC[-excludeMarkers]
markers_2h_FC.cluster_no0 <- markers_2h_FC.cluster[-excludeMarkers]
markersMatrix <- assay(sce.ifm.2h, "logcounts")[markers_2h_FC_no0,]
markersMatrix <- t(scale(t(markersMatrix)))
rownames(markersMatrix) <- with(
    rowData(sce.2h), gene_name[match(rownames(markersMatrix), gene_id)])
# maxMatrix <- max(abs(range(markersMatrix))) * ht.scale.percentile
ht_scale <- circlize::colorRamp2(c(-4, 0, 4), ht.col) # use maxMatrix?
for (clusterId in unique(markers_2h_FC.cluster_no0)){
  clusterMatrix <- markersMatrix[,sce.ifm.2h$quickCluster.2h == clusterId]
  ha_tmp <- HeatmapAnnotation(
    df = data.frame(subset(
      colData(sce.ifm.2h), quickCluster.2h == clusterId,
      c("Infection", "Status")
    )),
    col = list(
      Infection = col.infection,
      Status = col.status
    )
  )
  ht_list[[clusterId]] <- Heatmap(
    matrix = clusterMatrix,
    name = ifelse(clusterId == "1", "z-score", clusterId),
    col = ht_scale,
    row_names_gp = gpar(cex=0.4),
    show_heatmap_legend = clusterId == "1",
    show_row_names = clusterId == "1",
    cluster_rows = FALSE,
    show_column_names = FALSE, show_column_dend = FALSE,
    row_names_side = "left",
    top_annotation = ha_tmp,
    split = markers_2h_FC.cluster_no0
    )
}
ht_final <- ht_list[["1"]] + ht_list[["2"]] + ht_list[["3"]]
draw(ht_final)
```

```{r markers_FC_logcounts_2h_no0_pdf, include=FALSE}
pdf(
  file.path(outdir, "2h_markers_FC_heatmap_logcounts_no0.pdf"),
  width = 10, height = 15)
draw(ht_final)
dev.off()
```

```{r markers_FC_logcounts_2h_no0_annolink, include=FALSE}
ht_list <- list()
excludeMarkers <- which(markers_2h_FC.cluster == "0")
markers_2h_FC_no0 <- markers_2h_FC[-excludeMarkers]
markers_2h_FC.cluster_no0 <- markers_2h_FC.cluster[-excludeMarkers]
markersMatrix <- assay(sce.ifm.2h, "logcounts")[markers_2h_FC_no0,]
markersMatrix <- t(scale(t(markersMatrix)))
rownames(markersMatrix) <- with(
    rowData(sce.2h), gene_name[match(rownames(markersMatrix), gene_id)])
# maxMatrix <- max(abs(range(markersMatrix))) * ht.scale.percentile
ht_scale <- circlize::colorRamp2(c(-4, 0, 4), ht.col) # use maxMatrix?
for (clusterId in unique(markers_2h_FC.cluster_no0)){
  clusterMatrix <- markersMatrix[,sce.ifm.2h$quickCluster.2h == clusterId]
  ht_list[[clusterId]] <- Heatmap(
    matrix = clusterMatrix,
    name = ifelse(clusterId == "1", "z-score", clusterId),
    col = ht_scale,
    row_names_gp = gpar(cex=0.4),
    show_heatmap_legend = clusterId == "1",
    show_row_names = clusterId == "1",
    cluster_rows = FALSE,
    show_column_names = FALSE, show_column_dend = FALSE,
    row_names_side = "left",
    split = markers_2h_FC.cluster_no0
    )
}
geneNames <- c(
  "MMP19","CMKLR1","CCL24","CYBB","GSDMA","CTSL","RNASE1","LAMP1","CXCL2",
  "KLHL21","MMP9","SERPINE1","MAFB","SERPINB2","CD163","CCL7","SPSB1","IL10",
  "CCL26","CHST13","ME1","FAM83G","IRAK3","MAPK13","CHST15","TMEM158","LGMN",
  "TNF","CD40","CLEC10A","CLEC4A","GOLGA8B","ADAM19","GOLGA8A","TUBA1A","CD1C","CD1A",
  "SLC25A5","GALM"
)
ht_final <- ht_list[["1"]] + ht_list[["2"]] + ht_list[["3"]] + rowAnnotation(
  link = row_anno_link(
    at = which(rownames(markersMatrix) %in% geneNames),
    labels = rownames(markersMatrix)[which(rownames(markersMatrix) %in% geneNames)],
    labels_gp = gpar(cex=2/3),
    link_width = unit(0.5, "inches")
  ),
  width = unit(1, "cm") + max_text_width(labels)
)
pdf(
  file.path(outdir, "2h_markers_FC_heatmap_logcounts_no0_rowLink.pdf"),
  width = 10, height = 10)
draw(ht_final)
dev.off()
```

### 4h

```{r markers_FC_logcounts_4h, echo=FALSE, fig.height=10, message=FALSE}
ht_list <- list()
markersMatrix <- assay(sce.ifm.4h, "logcounts")[markers_4h_FC,]
markersMatrix <- t(scale(t(markersMatrix)))
rownames(markersMatrix) <- with(
    rowData(sce.4h), gene_name[match(rownames(markersMatrix), gene_id)])
# maxMatrix <- max(abs(range(markersMatrix))) * ht.scale.percentile
ht_scale <- circlize::colorRamp2(c(-4, 0, 4), ht.col) # use maxMatrix?
for (clusterId in levels(sce.ifm.4h$quickCluster.4h)){
  clusterMatrix <- markersMatrix[,sce.ifm.4h$quickCluster.4h == clusterId]
  ha_tmp <- HeatmapAnnotation(
    df = data.frame(subset(
      colData(sce.ifm.4h), quickCluster.4h == clusterId,
      c("Infection", "Status")
    )),
    col = list(
      Infection = col.infection,
      Status = col.status
    )
  )
  ht_list[[clusterId]] <- Heatmap(
    matrix = clusterMatrix,
    name = ifelse(
      clusterId == levels(sce.ifm.4h$quickCluster.4h)[1],
      "z-score",
      clusterId),
    col = ht_scale,
    row_names_gp = gpar(cex=0.4),
    show_heatmap_legend =
      clusterId == levels(sce.ifm.4h$quickCluster.4h)[1],
    show_row_names = clusterId == levels(sce.ifm.4h$quickCluster.4h)[1],
    cluster_rows = FALSE,
    show_column_names = FALSE,
    show_column_dend = FALSE,
    top_annotation = ha_tmp,
    row_names_side = "left",
    split = markers_4h_FC.cluster
    )
}
ht_final <- ht_list[["0"]] + ht_list[["1"]] + ht_list[["2"]] + ht_list[["3"]]
draw(ht_final)
```

```{r markers_FC_logcounts_4h_pdf, include=FALSE}
pdf(
  file.path(outdir, "4h_markers_FC_heatmap_logcounts.pdf"),
  width = 10, height = 15)
draw(ht_final)
dev.off()
```

### 6h

```{r markers_FC_logcounts_6h, echo=FALSE, fig.height=10, message=FALSE}
ht_list <- list()
markersMatrix <- assay(sce.ifm.6h, "logcounts")[markers_6h_FC,]
markersMatrix <- t(scale(t(markersMatrix)))
rownames(markersMatrix) <- with(
    rowData(sce.6h), gene_name[match(rownames(markersMatrix), gene_id)])
# maxMatrix <- max(abs(range(markersMatrix))) * ht.scale.percentile
ht_scale <- circlize::colorRamp2(c(-4, 0, 4), ht.col) # use maxMatrix?
for (clusterId in levels(sce.ifm.6h$quickCluster.6h)){
  clusterMatrix <- markersMatrix[,sce.ifm.6h$quickCluster.6h == clusterId]
  ha_tmp <- HeatmapAnnotation(
    df = data.frame(subset(
      colData(sce.ifm.6h), quickCluster.6h == clusterId,
      c("Infection", "Status")
    )),
    col = list(
      Infection = col.infection,
      Status = col.status
    )
  )
  ht_list[[clusterId]] <- Heatmap(
    matrix = clusterMatrix,
    name = ifelse(
      clusterId == levels(sce.ifm.6h$quickCluster.6h)[1],
      "z-score",
      clusterId),
    col = ht_scale,
    row_names_gp = gpar(cex=0.4),
    show_heatmap_legend =
      clusterId == levels(sce.ifm.6h$quickCluster.6h)[1],
    show_row_names = clusterId == levels(sce.ifm.6h$quickCluster.6h)[1],
    cluster_rows = FALSE,
    show_column_names = FALSE,
    show_column_dend = FALSE,
    row_names_side = "left",
    top_annotation = ha_tmp,
    split = markers_6h_FC.cluster
    )
}
ht_final <- ht_list[["0"]] + ht_list[["1"]] + ht_list[["2"]] + ht_list[["3"]]
draw(ht_final)
```

```{r markers_FC_logcounts_6h_pdf, include=FALSE}
pdf(
  file.path(outdir, "6h_markers_FC_heatmap_logcounts.pdf"),
  width = 10, height = 15)
draw(ht_final)
dev.off()
```

# Top 50 markers by significance

## Identification {.tabset}

For each time point, let us identify gene markers of each cluster,
and also trace the respective cluster for which each gene is a marker:

**Note**:
a gene may be identified as a marker for more than one cluster.

### 2h

```{r markers_2h_p}
markers_2h_p <- character()
markers_2h_p.cluster <- factor(NULL, levels(sce.ifm.2h$quickCluster.2h))
c2h <- grep("^2h", names(scde.res), value = TRUE)
for (contrastName in c2h){
  clusterId <- gsub("2h_cluster(.)-Others", "\\1", contrastName)
  scde.table <- addGeneName(convert.z.score(scde.res[[contrastName]]))
  scde.table <- subset(scde.table, mle > 0 & p.value < 0.01)
  scde.table <- scde.table[order(scde.table$p.value),]
  scde.markers <- head(rownames(scde.table), 100)
  markers_2h_p <- c(markers_2h_p, scde.markers)
  markers_2h_p.cluster <-
    c(markers_2h_p.cluster, rep(clusterId, length(scde.markers)))
}
```

### 4h

```{r markers_4h_p}
markers_4h_p <- character()
markers_4h_p.cluster <- factor(NULL, levels(sce.ifm.4h$quickCluster.4h))
c4h <- grep("^4h", names(scde.res), value = TRUE)
for (contrastName in c4h){
  clusterId <- gsub("4h_cluster(.)-Others", "\\1", contrastName)
  scde.table <- addGeneName(convert.z.score(scde.res[[contrastName]]))
  scde.table <- subset(scde.table, mle > 0 & p.value < 0.01)
  scde.table <- scde.table[order(scde.table$p.value),]
  scde.markers <- head(rownames(scde.table), 50)
  markers_4h_p <- c(markers_4h_p, scde.markers)
  markers_4h_p.cluster <-
    c(markers_4h_p.cluster, rep(clusterId, length(scde.markers)))
}
```

### 6h

```{r markers_6h_p}
markers_6h_p <- character()
markers_6h_p.cluster <- factor(NULL, levels(sce.ifm.6h$quickCluster.6h))
c6h <- grep("^6h", names(scde.res), value = TRUE)
for (contrastName in c6h){
  clusterId <- gsub("6h_cluster(.)-Others", "\\1", contrastName)
  scde.table <- addGeneName(convert.z.score(scde.res[[contrastName]]))
  scde.table <- subset(scde.table, mle > 0 & p.value < 0.01)
  scde.table <- scde.table[order(scde.table$p.value),]
  scde.markers <- head(rownames(scde.table), 50)
  markers_6h_p <- c(markers_6h_p, scde.markers)
  markers_6h_p.cluster <-
    c(markers_6h_p.cluster, rep(clusterId, length(scde.markers)))
}
```

## Heat map of fold-change for (up to) 50 markers {.tabset}

For each time point, let us display up to 50 marker genes for each cluster, selected for having:

* *P*-value less than 0.01
* *mle* of log fold-change among the 50 largest positive values for a cluster

### 2h

```{r markers_P_2h, echo=FALSE, fig.height=10, message=FALSE}
markersMatrix <- matrix(
  nrow = length(markers_2h_p),
  ncol = length(c2h),
  dimnames = list(
    gene_id = markers_2h_p,
    cluster = gsub("2h_cluster(.)-Others", "\\1", c2h)
  ))
for (contrastName in c2h){
  scde.table <- addGeneName(convert.z.score(scde.res[[contrastName]]))
  clusterName <- gsub("2h_cluster(.)-Others", "\\1", contrastName)
  markersMatrix[,clusterName] <- scde.table[rownames(markersMatrix),"mle"]
}
rownames(markersMatrix) <-
  with(rowData(sce.2h), gene_name[match(rownames(markersMatrix), gene_id)])
ht <- Heatmap(
  matrix = markersMatrix, name = "log2FC",
  cluster_rows = FALSE, cluster_columns = FALSE,
  row_names_gp = gpar(cex=0.4),
  split = markers_2h_p.cluster
  )
draw(ht)
```

```{r markers_P_2h_pdf, include=FALSE}
pdf(file.path(outdir, "2h_markers_P_heatmap_FC.pdf"), width = 8, height = 15)
draw(ht)
dev.off()
```

### 4h

```{r markers_P_4h, echo=FALSE, fig.height=10}
markersMatrix <- matrix(
  nrow = length(markers_4h_p),
  ncol = length(c4h),
  dimnames = list(
    gene_id = markers_4h_p,
    cluster = gsub("4h_cluster(.)-Others", "\\1", c4h)
  ))
for (contrastName in c4h){
  scde.table <- addGeneName(convert.z.score(scde.res[[contrastName]]))
  clusterName <- gsub("4h_cluster(.)-Others", "\\1", contrastName)
  markersMatrix[,clusterName] <- scde.table[rownames(markersMatrix),"mle"]
}
rownames(markersMatrix) <-
  with(rowData(sce.4h), gene_name[match(rownames(markersMatrix), gene_id)])
ht <- Heatmap(
  matrix = markersMatrix, name = "log2FC",
  cluster_rows = FALSE, cluster_columns = FALSE,
  row_names_gp = gpar(cex=0.4),
  split = markers_4h_p.cluster
  )
draw(ht)
```

```{r markers_P_4h_pdf, include=FALSE}
pdf(file.path(outdir, "4h_markers_P_heatmap_FC.pdf"), width = 8, height = 15)
draw(ht)
dev.off()
```

### 6h

```{r markers_P_6h, echo=FALSE, fig.height=10}
markersMatrix <- matrix(
  nrow = length(markers_6h_p),
  ncol = length(c6h),
  dimnames = list(
    gene_id = markers_6h_p,
    cluster = gsub("6h_cluster(.)-Others", "\\1", c6h)
  ))
for (contrastName in c6h){
  scde.table <- addGeneName(convert.z.score(scde.res[[contrastName]]))
  clusterName <- gsub("6h_cluster(.)-Others", "\\1", contrastName)
  markersMatrix[,clusterName] <- scde.table[rownames(markersMatrix),"mle"]
}
rownames(markersMatrix) <-
  with(rowData(sce.6h), gene_name[match(rownames(markersMatrix), gene_id)])
ht <- Heatmap(
  matrix = markersMatrix, name = "log2FC",
  cluster_rows = FALSE, cluster_columns = FALSE,
  row_names_gp = gpar(cex=0.4),
  split = markers_6h_p.cluster
  )
draw(ht)
```

```{r markers_P_6h_pdf, include=FALSE}
pdf(file.path(outdir, "6h_markers_P_heatmap_FC.pdf"), width = 8, height = 15)
draw(ht)
dev.off()
```

## Heat map of row-scaled log-counts for (up to) 50 markers {.tabset}

Notably, let us use here the same markers as defined in the previous section,
namely:

* *P*-value less than 0.01
* *P*-value among the 50 smallest values for a cluster

### 2h

```{r markers_p_logcounts_2h, echo=FALSE, fig.height=10, message=FALSE}
ht_list <- list()
markersMatrix <- assay(sce.ifm.2h, "logcounts")[markers_2h_p,]
markersMatrix <- t(scale(t(markersMatrix)))
rownames(markersMatrix) <- with(
    rowData(sce.2h), gene_name[match(rownames(markersMatrix), gene_id)])
# maxMatrix <- max(abs(range(markersMatrix))) * ht.scale.percentile
ht_scale <- circlize::colorRamp2(c(-4, 0, 4), ht.col) # use maxMatrix?
for (clusterId in levels(sce.ifm.2h$quickCluster.2h)){
  clusterMatrix <- markersMatrix[,sce.ifm.2h$quickCluster.2h == clusterId]
  ha_tmp <- HeatmapAnnotation(
    df = data.frame(subset(
      colData(sce.ifm.2h), quickCluster.2h == clusterId,
      c("Infection", "Status")
    )),
    col = list(
      Infection = col.infection,
      Status = col.status
    )
  )
  ht_list[[clusterId]] <- Heatmap(
    matrix = clusterMatrix,
    name = ifelse(
      clusterId == levels(sce.ifm.2h$quickCluster.2h)[1],
      "z-score",
      clusterId),
    col = ht_scale,
    row_names_gp = gpar(cex=0.4),
    show_heatmap_legend =
      clusterId == levels(sce.ifm.2h$quickCluster.2h)[1],
    show_row_names = clusterId == levels(sce.ifm.2h$quickCluster.2h)[1],
    cluster_rows = FALSE,
    show_column_names = FALSE, show_column_dend = FALSE,
    row_names_side = "left",
    top_annotation = ha_tmp,
    split = markers_2h_p.cluster
    )
}
ht_final <- ht_list[["0"]] + ht_list[["1"]] + ht_list[["2"]] + ht_list[["3"]]
draw(ht_final)
```

```{r markers_p_logcounts_2h_pdf, include=FALSE}
pdf(
  file.path(outdir, "2h_markers_P_heatmap_logcounts.pdf"),
  width = 10, height = 15)
draw(ht_final)
dev.off()
```

### 2h (w/out '0')

```{r markers_p_logcounts_2h_no0, echo=FALSE, fig.height=10, message=FALSE}
ht_list <- list()
excludeMarkers <- which(markers_2h_p.cluster == "0")
markers_2h_p_no0 <- markers_2h_p[-excludeMarkers]
markers_2h_p.cluster_no0 <- markers_2h_p.cluster[-excludeMarkers]
markersMatrix <- assay(sce.ifm.2h, "logcounts")[markers_2h_p_no0,]
markersMatrix <- t(scale(t(markersMatrix)))
rownames(markersMatrix) <- with(
    rowData(sce.2h), gene_name[match(rownames(markersMatrix), gene_id)])
# maxMatrix <- max(abs(range(markersMatrix))) * ht.scale.percentile
ht_scale <- circlize::colorRamp2(c(-4, 0, 4), ht.col) # use maxMatrix?
for (clusterId in unique(markers_2h_p.cluster_no0)){
  clusterMatrix <- markersMatrix[,sce.ifm.2h$quickCluster.2h == clusterId]
  ha_tmp <- HeatmapAnnotation(
    df = data.frame(subset(
      colData(sce.ifm.2h), quickCluster.2h == clusterId,
      c("Infection", "Status")
    )),
    col = list(
      Infection = col.infection,
      Status = col.status
    )
  )
  ht_list[[clusterId]] <- Heatmap(
    matrix = clusterMatrix,
    name = ifelse(clusterId == "1", "z-score", clusterId),
    col = ht_scale,
    row_names_gp = gpar(cex=0.4),
    show_heatmap_legend = clusterId == "1",
    show_row_names = clusterId == "1",
    cluster_rows = FALSE,
    show_column_names = FALSE, show_column_dend = FALSE,
    row_names_side = "left",
    top_annotation = ha_tmp,
    split = markers_2h_p.cluster_no0
    )
}
ht_final <- ht_list[["1"]] + ht_list[["2"]] + ht_list[["3"]]
draw(ht_final)
```

```{r markers_p_logcounts_2h_no0_pdf, include=FALSE}
pdf(
  file.path(outdir, "2h_markers_P_heatmap_logcounts_no0.pdf"),
  width = 10, height = 15)
draw(ht_final)
dev.off()
```

```{r markers_p_logcounts_2h_no0_annolink, include=FALSE}
ht_list <- list()
excludeMarkers <- which(markers_2h_p.cluster == "0")
markers_2h_p_no0 <- markers_2h_p[-excludeMarkers]
markers_2h_p.cluster_no0 <- markers_2h_p.cluster[-excludeMarkers]
markersMatrix <- assay(sce.ifm.2h, "logcounts")[markers_2h_p_no0,]
markersMatrix <- t(scale(t(markersMatrix)))
rownames(markersMatrix) <- with(
    rowData(sce.2h), gene_name[match(rownames(markersMatrix), gene_id)])
# maxMatrix <- max(abs(range(markersMatrix))) * ht.scale.percentile
ht_scale <- circlize::colorRamp2(c(-4, 0, 4), ht.col) # use maxMatrix?
for (clusterId in unique(markers_2h_p.cluster_no0)){
  clusterMatrix <- markersMatrix[,sce.ifm.2h$quickCluster.2h == clusterId]
  ha_tmp <- HeatmapAnnotation(
    df = data.frame(subset(
      colData(sce.ifm.2h), quickCluster.2h == clusterId,
      c("Infection", "Status")
    )),
    col = list(
      Infection = col.infection,
      Status = col.status
    )
  )
  ht_list[[clusterId]] <- Heatmap(
    matrix = clusterMatrix,
    name = ifelse(clusterId == "1", "z-score", clusterId),
    col = ht_scale,
    row_names_gp = gpar(cex=0.4),
    show_heatmap_legend = clusterId == "1",
    show_row_names = clusterId == "1",
    cluster_rows = FALSE,
    show_column_names = FALSE, show_column_dend = FALSE,
    row_names_side = "left",
    top_annotation = ha_tmp,
    split = markers_2h_p.cluster_no0
    )
}
geneNames <- c(
  "MMP19","CMKLR1","CCL24","CYBB","GSDMA","CTSL","RNASE1","LAMP1","CXCL2",
  "KLHL21","MMP9","SERPINE1","MAFB","SERPINB2","CD163","CCL7","SPSB1","IL10",
  "CCL26","CHST13","ME1","FAM83G","IRAK3","MAPK13","CHST15","TMEM158","LGMN",
  "TNF","CD40","CLEC10A","CLEC4A","GOLGA8B","ADAM19","GOLGA8A","TUBA1A","CD1C","CD1A",
  "SLC25A5","GALM"
)
ht_final <- ht_list[["1"]] + ht_list[["2"]] + ht_list[["3"]] + rowAnnotation(
  link = row_anno_link(
    at = which(rownames(markersMatrix) %in% geneNames),
    labels = rownames(markersMatrix)[which(rownames(markersMatrix) %in% geneNames)],
    labels_gp = gpar(cex=2/3),
    link_width = unit(0.5, "inches")
  ),
  width = unit(1, "cm") + max_text_width(labels)
)
pdf(
  file.path(outdir, "2h_markers_P_heatmap_logcounts_no0_rowLink.pdf"),
  width = 10, height = 10)
draw(ht_final)
dev.off()
```

### 4h

```{r markers_p_logcounts_4h, echo=FALSE, fig.height=10, message=FALSE}
ht_list <- list()
markersMatrix <- assay(sce.ifm.4h, "logcounts")[markers_4h_p,]
markersMatrix <- t(scale(t(markersMatrix)))
rownames(markersMatrix) <- with(
    rowData(sce.4h), gene_name[match(rownames(markersMatrix), gene_id)])
# maxMatrix <- max(abs(range(markersMatrix))) * ht.scale.percentile
ht_scale <- circlize::colorRamp2(c(-4, 0, 4), ht.col) # use maxMatrix?
for (clusterId in levels(sce.ifm.4h$quickCluster.4h)){
  clusterMatrix <- markersMatrix[,sce.ifm.4h$quickCluster.4h == clusterId]
  ha_tmp <- HeatmapAnnotation(
    df = data.frame(subset(
      colData(sce.ifm.4h), quickCluster.4h == clusterId,
      c("Infection", "Status")
    )),
    col = list(
      Infection = col.infection,
      Status = col.status
    )
  )
  ht_list[[clusterId]] <- Heatmap(
    matrix = clusterMatrix,
    name = ifelse(
      clusterId == levels(sce.ifm.4h$quickCluster.4h)[1],
      "z-score",
      clusterId),
    col = ht_scale,
    row_names_gp = gpar(cex=0.4),
    show_heatmap_legend =
      clusterId == levels(sce.ifm.4h$quickCluster.4h)[1],
    show_row_names = clusterId == levels(sce.ifm.4h$quickCluster.4h)[1],
    cluster_rows = FALSE,
    show_column_names = FALSE, show_column_dend = FALSE,
    row_names_side = "left",
    top_annotation = ha_tmp,
    split = markers_4h_p.cluster
    )
}
ht_final <- ht_list[["0"]] + ht_list[["1"]] + ht_list[["2"]] + ht_list[["3"]]
draw(ht_final)
```

```{r markers_p_logcounts_4h_pdf, include=FALSE}
pdf(
  file.path(outdir, "4h_markers_P_heatmap_logcounts.pdf"),
  width = 10, height = 15)
draw(ht_final)
dev.off()
```

```{r markers_p_logcounts_4h_no0_annolink, include=FALSE}
ht_list <- list()
markersMatrix <- assay(sce.ifm.4h, "logcounts")[markers_4h_p,]
markersMatrix <- t(scale(t(markersMatrix)))
rownames(markersMatrix) <- with(
    rowData(sce.4h), gene_name[match(rownames(markersMatrix), gene_id)])
# maxMatrix <- max(abs(range(markersMatrix))) * ht.scale.percentile
ht_scale <- circlize::colorRamp2(c(-4, 0, 4), ht.col) # use maxMatrix?
for (clusterId in unique(markers_4h_p.cluster)){
  clusterMatrix <- markersMatrix[,sce.ifm.4h$quickCluster.4h == clusterId]
  ha_tmp <- HeatmapAnnotation(
    df = data.frame(subset(
      colData(sce.ifm.4h), quickCluster.4h == clusterId,
      c("Infection", "Status")
    )),
    col = list(
      Infection = col.infection,
      Status = col.status
    )
  )
  ht_list[[clusterId]] <- Heatmap(
    matrix = clusterMatrix,
    name = ifelse(clusterId == "1", "z-score", clusterId),
    col = ht_scale,
    row_names_gp = gpar(cex=0.4),
    show_heatmap_legend = clusterId == "1",
    show_row_names = clusterId == "1",
    cluster_rows = FALSE,
    show_column_names = FALSE, show_column_dend = FALSE,
    row_names_side = "left",
    top_annotation = ha_tmp,
    split = markers_4h_p.cluster
    )
}
geneNames <- c(
  "CTSL","CTSD","S100A9","CCL7","CTSS","CCL24",
  "CD40","IL1B","CCL1","STAT5A","CD83","TNF",
  "IL1A","CEBPA"
)
ht_final <- ht_list[["1"]] + ht_list[["2"]] + ht_list[["3"]] + rowAnnotation(
  link = row_anno_link(
    at = which(rownames(markersMatrix) %in% geneNames),
    labels = rownames(markersMatrix)[which(rownames(markersMatrix) %in% geneNames)],
    labels_gp = gpar(cex=2/3),
    link_width = unit(0.5, "inches")
  ),
  width = unit(1, "cm") + max_text_width(labels)
)
pdf(
  file.path(outdir, "4h_markers_P_heatmap_logcounts_no0_rowLink.pdf"),
  width = 10, height = 10)
draw(ht_final)
dev.off()
```

### 6h

```{r markers_p_logcounts_6h, echo=FALSE, fig.height=10, message=FALSE}
ht_list <- list()
markersMatrix <- assay(sce.ifm.6h, "logcounts")[markers_6h_p,]
markersMatrix <- t(scale(t(markersMatrix)))
rownames(markersMatrix) <- with(
    rowData(sce.6h), gene_name[match(rownames(markersMatrix), gene_id)])
# maxMatrix <- max(abs(range(markersMatrix))) * ht.scale.percentile
ht_scale <- circlize::colorRamp2(c(-4, 0, 4), ht.col) # use maxMatrix?
for (clusterId in levels(sce.ifm.6h$quickCluster.6h)){
  clusterMatrix <- markersMatrix[,sce.ifm.6h$quickCluster.6h == clusterId]
  ha_tmp <- HeatmapAnnotation(
    df = data.frame(subset(
      colData(sce.ifm.6h), quickCluster.6h == clusterId,
      c("Infection", "Status")
    )),
    col = list(
      Infection = col.infection,
      Status = col.status
    )
  )
  ht_list[[clusterId]] <- Heatmap(
    matrix = clusterMatrix,
    name = ifelse(
      clusterId == levels(sce.ifm.6h$quickCluster.6h)[1],
      "z-score",
      clusterId),
    col = ht_scale,
    row_names_gp = gpar(cex=0.4),
    show_heatmap_legend =
      clusterId == levels(sce.ifm.6h$quickCluster.6h)[1],
    show_row_names = clusterId == levels(sce.ifm.6h$quickCluster.6h)[1],
    cluster_rows = FALSE,
    show_column_names = FALSE, show_column_dend = FALSE,
    row_names_side = "left",
    top_annotation = ha_tmp,
    split = markers_6h_p.cluster
    )
}
ht_final <- ht_list[["0"]] + ht_list[["1"]] + ht_list[["2"]] + ht_list[["3"]]
draw(ht_final)
```

```{r markers_p_logcounts_6h_pdf, include=FALSE}
pdf(
  file.path(outdir, "6h_markers_P_heatmap_logcounts.pdf"),
  width = 10, height = 15)
draw(ht_final)
dev.off()
```

```{r markers_p_logcounts_6h_no0_annolink, include=FALSE}
ht_list <- list()
markersMatrix <- assay(sce.ifm.6h, "logcounts")[markers_6h_p,]
markersMatrix <- t(scale(t(markersMatrix)))
rownames(markersMatrix) <- with(
    rowData(sce.6h), gene_name[match(rownames(markersMatrix), gene_id)])
# maxMatrix <- max(abs(range(markersMatrix))) * ht.scale.percentile
ht_scale <- circlize::colorRamp2(c(-4, 0, 4), ht.col) # use maxMatrix?
for (clusterId in unique(markers_6h_p.cluster)){
  clusterMatrix <- markersMatrix[,sce.ifm.6h$quickCluster.6h == clusterId]
  ha_tmp <- HeatmapAnnotation(
    df = data.frame(subset(
      colData(sce.ifm.6h), quickCluster.6h == clusterId,
      c("Infection", "Status")
    )),
    col = list(
      Infection = col.infection,
      Status = col.status
    )
  )
  ht_list[[clusterId]] <- Heatmap(
    matrix = clusterMatrix,
    name = ifelse(clusterId == "1", "z-score", clusterId),
    col = ht_scale,
    row_names_gp = gpar(cex=0.4),
    show_heatmap_legend = clusterId == "1",
    show_row_names = clusterId == "1",
    cluster_rows = FALSE,
    show_column_names = FALSE, show_column_dend = FALSE,
    row_names_side = "left",
    top_annotation = ha_tmp,
    split = markers_6h_p.cluster
    )
}
geneNames <- c(
  "IL1A","IL1B","IL12B","IL23A","IL6","CTSL","DEFB","MARCH1","BLOC1S3","RNF139",
  "ATF3","CLECL1"
)
ht_final <- ht_list[["1"]] + ht_list[["2"]] + ht_list[["3"]] + rowAnnotation(
  link = row_anno_link(
    at = which(rownames(markersMatrix) %in% geneNames),
    labels = rownames(markersMatrix)[which(rownames(markersMatrix) %in% geneNames)],
    labels_gp = gpar(cex=2/3),
    link_width = unit(0.5, "inches")
  ),
  width = unit(1, "cm") + max_text_width(labels)
)
pdf(
  file.path(outdir, "6h_markers_P_heatmap_logcounts_no0_rowLink.pdf"),
  width = 10, height = 10)
draw(ht_final)
dev.off()
```

# Gene ontology {.tabset}

Let us use the [goseq](http://bioconductor.org/packages/goseq) package to
identify the most enriched gene ontologies among the various lists of positive markers.
Note that we restrict the results to GO categories associated with at least
**10** genes, for robustness.

Let us first define:

* a list to store the GO enrichment tables returned by *goseq*

```{r goseq.res_init, eval=FALSE}
goseq.res <- list()
```

* the gene length information reported by *featureCounts*:

```{r geneLengths}
geneLengths <- width(sce.endo)
names(geneLengths) <- rownames(sce.endo)
```

## Computation {.tabset}

### 2h

```{r goseq_2h_p, eval=FALSE}
bg.2h <- rownames(cd.ifm.2h)
for (contrastName in c2h){
  message(contrastName)
  clusterId <- gsub("2h_cluster(.)-Others", "\\1", contrastName)
  scde.table <- convert.z.score(scde.res[[contrastName]])
  scde.table <- subset(scde.table, mle > 0 & p.value < 0.01)
  marker.genes <- (bg.2h %in% rownames(scde.table))
  names(marker.genes) <- bg.2h; table(marker.genes)
  pwf <- nullp(marker.genes, bias.data = geneLengths[names(marker.genes)])
  go.res <- goseq(pwf, "hg38", "ensGene")
  goseq.res[[contrastName]] <- go.res
  saveRDS(goseq.res, "rds/goseq_markers.rds") # checkpoint
}
```

### 4h

```{r goseq_4h_p, eval=FALSE}
bg.4h <- rownames(cd.ifm.4h)
for (contrastName in c4h){
  message(contrastName)
  clusterId <- gsub("4h_cluster(.)-Others", "\\1", contrastName)
  scde.table <- convert.z.score(scde.res[[contrastName]])
  scde.table <- subset(scde.table, mle > 0 & p.value < 0.01)
  marker.genes <- (bg.4h %in% rownames(scde.table))
  names(marker.genes) <- bg.4h; table(marker.genes)
  pwf <- nullp(marker.genes, bias.data = geneLengths[names(marker.genes)])
  go.res <- goseq(pwf, "hg38", "ensGene")
  goseq.res[[contrastName]] <- go.res
  saveRDS(goseq.res, "rds/goseq_markers.rds") # checkpoint
}
```

### 6h

```{r goseq_6h_p, eval=FALSE}
bg.6h <- rownames(cd.ifm.6h)
for (contrastName in c6h){
  message(contrastName)
  clusterId <- gsub("6h_cluster(.)-Others", "\\1", contrastName)
  scde.table <- convert.z.score(scde.res[[contrastName]])
  scde.table <- subset(scde.table, mle > 0 & p.value < 0.01)
  marker.genes <- (bg.6h %in% rownames(scde.table))
  names(marker.genes) <- bg.6h; table(marker.genes)
  pwf <- nullp(marker.genes, bias.data = geneLengths[names(marker.genes)])
  go.res <- goseq(pwf, "hg38", "ensGene")
  goseq.res[[contrastName]] <- go.res
  saveRDS(goseq.res, "rds/goseq_markers.rds") # checkpoint
}
```

```{r goseq.table_csv, echo=FALSE, eval=FALSE}
for (contrastName in names(goseq.res)){
  goseq.table <- goseq.res[[contrastName]]
  write.csv(
    goseq.table,
    file.path(outdir, sprintf("SCDE_GO_%s.csv", contrastName)))
}
```
